pyproject.toml
[project]
name = "nonebot_plugin_poke"
dynamic = ["version"]
description = "自定义群聊戳戳事件 plugin for NoneBot2"
authors = [{ name = "Agnes_Digital", email = "Z735803792@163.com" }]
requires-python = ">=3.9,<4.0"
readme = "README.md"
license = { text = "MIT" }

homepage = "https://github.com/Agnes4m/nonebot_plugin_poke"
repository = "https://github.com/Agnes4m/nonebot_plugin_poke"
keywords = ["nonebot2", "plugin", "poke"]

classifiers = [
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Operating System :: OS Independent",
]
include = ["LICENSE", "README.md"]

dependencies = [
    "nonebot2>=2.1.0",
    "nonebot-plugin-alconna>=0.58.5",
    "nonebot-adapter-onebot>=2.1.5",
    "aiofiles>=23.1.0",
    "aiohttp>=3.8.5",
    "pydantic>=2.1.1",
    "pillow>=9.0.0",
]

[tool.pdm.build]
includes = []

[tool.pdm.version]
source = "file"
path = "nonebot_plugin_poke/__init__.py"

[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"


[tool.black]
line-length = 89
target-version = ["py39", "py310", "py311", "py312", "py313"]
include = '\.pyi?$'
# skip-string-normalization = true

[tool.basedpyright]
pythonVersion = "3.9"
defineConstant = { PYDANTIC_V2 = true }
typeCheckingMode = "standard"
reportShadowedImports = false
exclude = ["**/node_modules", "**/.venv", "./nonebot_plugin_poke"]
executionEnvironments = [
    { root = "typings", reportIncompatibleMethodOverride = false, reportSelfClsParameterName = false },
    { root = "." },
]


[tool.ruff.lint.isort]
extra-standard-library = ["typing_extensions"]

[tool.ruff.lint]
ignore = [
    "B008",
    "B905",
    "E402",
    "E501",
    "FBT001",
    "FBT002",
    "FBT003",
    "PGH003",
    "RUF001",
    "RUF002",
    "RUF003",
    "RUF006",
    "RUF100",
    "TRY002",
    "TRY003",

]
select = [
    "A",
    "ARG",
    "ASYNC",
    "B",
    "C4",
    "COM",
    "E",
    "F",
    "FBT",
    "FLY",
    "I",
    "ISC",
    "N",
    "PIE",
    "PGH",
    "PTH",
    "PYI",
    "Q",
    "RET",
    "RSE",
    "RUF",
    "SIM",
    "SLF",
    "SLOT",
    "TRY",
]
